requirements.txt
# =====================================================
# RENDEREXPO AI STUDIO - Backend Requirements (Windows)
# FINAL STABLE VERSION (NO CONFLICTS)
# =====================================================

# --- Web framework ---
fastapi==0.109.2
uvicorn[standard]==0.29.0
starlette==0.36.3

# --- Core types & utils ---
pydantic==2.12.4
typing-extensions==4.14.1
tqdm==4.67.1
python-dotenv==1.0.1

# =====================================================
#  ML STACK (WINDOWS-COMPATIBLE)
# =====================================================
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# =====================================================
#  Diffusion Ecosystem
# =====================================================
diffusers==0.27.2
transformers==4.39.3
accelerate==0.28.0
safetensors==0.4.2
huggingface_hub==0.23.0

# =====================================================
#  Image / CV Essentials
# =====================================================
Pillow==10.2.0
opencv-python==4.9.0.80

# =====================================================
#  Extra utilities (MiDaS, ControlNet, ESRGAN depend on these)
# =====================================================
einops==0.7.0
timm==0.9.16
